增加外网支持

+++ sei-agent/src/main/java/com/changhong/sei/log/service/ScheduleService.java
@@ -14,7 +14,9 @@ import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
+import java.util.Objects;
 import java.util.Set;
+import java.util.concurrent.atomic.LongAdder;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -35,6 +37,8 @@ public class ScheduleService {
     private String env;
     @Autowired
     private ElasticSearchService elasticService;
+    // @Autowired
+    // private DiscoveryClient discoveryClient;
 
     // 创建 Pattern 对象
     private static final Pattern PATTERN = Pattern.compile("\\d{4}\\.\\d{1,2}\\.\\d{1,2}");
@@ -52,24 +56,27 @@ public class ScheduleService {
         if (resultData.successful()) {
             LocalDate targetData = LocalDate.now().minusDays(cleanupDays);
 
-            int count = 0;
+            LongAdder longAdder = new LongAdder();
             Set<String> set = resultData.getData();
-            for (String index : set) {
-                if (StringUtils.containsIgnoreCase(index, "-".concat(env).concat("-"))) {
-                    // 现在创建 matcher 对象
-                    Matcher m = PATTERN.matcher(index);
-                    if (m.find()) {
-                        String dateStr = m.group(0);
-                        LocalDate localDate = LocalDate.parse(dateStr, FMT);
-                        if (localDate.isBefore(targetData)) {
-                            elasticService.deleteIndex(index);
-                            count++;
-                            LogUtil.bizLog("成功清理{}的日志数据", index);
+            if (Objects.nonNull(set)) {
+                set.forEach(index -> {
+                    // 只删除符合应用日志命名规范的索引. 命名规则: 应用代码-环境代码-日期-fromkafka 如: sei-notify-dev-2022.02.23-fromkafka
+                    if (StringUtils.containsIgnoreCase(index, "-".concat(env).concat("-"))) {
+                        // 现在创建 matcher 对象
+                        Matcher m = PATTERN.matcher(index);
+                        if (m.find()) {
+                            String dateStr = m.group(0);
+                            LocalDate localDate = LocalDate.parse(dateStr, FMT);
+                            if (localDate.isBefore(targetData)) {
+                                elasticService.deleteIndex(index);
+                                longAdder.increment();
+                                LogUtil.bizLog("成功清理{}的日志数据", index);
+                            }
                         }
                     }
-                }
+                });
             }
-            LogUtil.bizLog("完成{}个日志清理", count);
+            LogUtil.bizLog("完成{}个日志清理", longAdder.sum());
         } else {
             LOG.error("定时清理日志出错: {}", resultData.getMessage());
         }
